Fix initial layer 3 Y position being off by a bit
I made some mistakes and also it turns out that the documentation is probably outdated. It says that initial layer 3 y position is `TTttttYYYYy` with 
```
SP3 	tttt----
SP2 	TT??----
LG4 	yOIB----
LG3 	YYYY----
```
but it actually seems to be now `TTTTttttYYYYy` with
```
SP3 	tttt----
SP2 	TTTT----
LG4 	yOIB----
LG3 	YYYY----
```
so the two unused bytes seem to now be part of the high part of the initial layer 3 Y position.